pyproject.toml
[tool.poetry]
name = "crb-inventory"
version = "0.1.0"
description = ""
authors = ["ph-cardoso <phlc.cardoso@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/ph-cardoso/crb-inventory"
repository = "https://github.com/ph-cardoso/crb-inventory"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
uuid-utils = "^0.9.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.10"
taskipy = "^1.13.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
testcontainers = "^4.7.0"
factory-boy = "^3.3.0"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.lint.per-file-ignores]
"**/{tests}/*" = ["E501"]

[tool.ruff.format]
preview = true

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
dev = 'fastapi dev crb_inventory/main.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=crb_inventory -vv'
post_test = 'coverage html'
local-env-up = 'docker compose --env-file .env -f compose.yml up -d'
local-env-down = 'docker compose --env-file .env -f compose.yml down'
clean = "rm -rf .coverage htmlcov .ruff_cache **/__pycache__"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
